# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Smart House API
  version: 0.1.0
  description: Public HTTP contract for the Smart House platform.
servers:
  - url: http://localhost:4000/api
paths:
  /health:
    get:
      operationId: getHealth
      summary: Read API health
      responses:
        "200":
          description: API is available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /spike/telemetry/latest:
    get:
      operationId: getLatestSpikeTelemetry
      summary: Read the latest persisted telemetry for the local spike
      responses:
        "200":
          description: Latest metric for every simulated device and metric pair.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LatestTelemetry"
  /organizations/{organizationId}/devices:
    get:
      operationId: listOrganizationDevices
      summary: List organization devices with inventory filters
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          description: Case-insensitive name or external ID search.
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          schema:
            type: string
            enum: [inactive, offline, online, disabled]
        - name: type
          in: query
          schema:
            type: string
            maxLength: 100
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of devices ordered by name, external ID, and ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceList"
  /organizations/{organizationId}/devices/{deviceId}:
    get:
      operationId: getOrganizationDevice
      summary: Read an authorized organization device
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: The device belongs to an organization the caller can access.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /organizations/{organizationId}/devices/{deviceId}/telemetry:
    get:
      operationId: getOrganizationDeviceTelemetry
      summary: Read historical telemetry for an authorized organization device
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/DeviceId"
        - name: metric
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: resolution
          in: query
          required: true
          schema:
            type: string
            enum: [auto, raw, 5m, 1h]
      responses:
        "200":
          description: Ordered raw or aggregated points. Relay aggregates use the last state, not an average.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceTelemetry"
  /organizations/{organizationId}/devices/{deviceId}/telemetry/export.csv:
    get:
      operationId: exportOrganizationDeviceTelemetryCsv
      summary: Export bounded historical telemetry as CSV for an authorized organization device
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - $ref: "#/components/parameters/DeviceId"
        - name: metric
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: resolution
          in: query
          required: true
          schema:
            type: string
            enum: [auto, raw, 5m, 1h]
      responses:
        "200":
          description: A CSV attachment, limited to 10,000 rows and export-specific maximum ranges.
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Attachment filename.
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /organizations/{organizationId}/alert-rules:
    get:
      operationId: listOrganizationAlertRules
      summary: List alert rules for an authorized organization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Threshold alert rules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRule"
    post:
      operationId: createOrganizationAlertRule
      summary: Create an alert rule for an organization device
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRule"
      responses:
        "201":
          description: Threshold alert rule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
  /organizations/{organizationId}/alerts:
    get:
      operationId: listOrganizationAlerts
      summary: List alerts for an authorized organization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - name: state
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved, silenced]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        "200":
          description: Alerts ordered by opening time.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /organizations/{organizationId}/alerts/{alertId}/{action}:
    post:
      operationId: transitionOrganizationAlert
      summary: Acknowledge, resolve, or silence an alert
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [acknowledge, resolve, silence]
      responses:
        "200":
          description: Alert state transition recorded and queued for realtime delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /organizations/{organizationId}/notifications:
    get:
      operationId: listOrganizationNotifications
      summary: List the authenticated member's alert notification inbox
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Authorized member inbox and unread count.
          content:
            application/json:
              schema:
                type: object
                required: [items, unreadCount]
                properties:
                  unreadCount: { type: integer, minimum: 0 }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/InboxNotification" }
  /organizations/{organizationId}/notifications/{notificationId}/read:
    post:
      operationId: markOrganizationNotificationRead
      summary: Mark an authenticated member's notification as read
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
        - name: notificationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Notification marked read.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxNotification" }
components:
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    HealthResponse:
      type: object
      required: [status, service, timestamp]
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
          const: api
    InboxNotification:
      type: object
      required: [id, alertId, title, body, severity, data, readAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        alertId: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        data: { type: object, additionalProperties: true }
        readAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [statusCode, message, path, timestamp]
      properties:
        statusCode:
          type: integer
          minimum: 400
          maximum: 599
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
    LatestTelemetry:
      type: object
      required: [correlationId, deviceId, metric, value, unit, occurredAt]
      properties:
        correlationId:
          type: string
        deviceId:
          type: string
        metric:
          type: string
        value:
          type: number
        unit:
          type: string
        occurredAt:
          type: string
          format: date-time
    Device:
      type: object
      required:
        [
          id,
          externalId,
          name,
          type,
          capabilityVersion,
          status,
          lastSeenAt,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        name:
          type: string
        type:
          type: string
        capabilityVersion:
          type: string
        status:
          type: string
          enum: [inactive, offline, online, disabled]
        lastSeenAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceList:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        nextCursor:
          type: [string, "null"]
    TelemetryPoint:
      type: object
      required: [occurredAt, value, unit]
      properties:
        occurredAt:
          type: string
          format: date-time
        value:
          type: number
        unit:
          type: string
    DeviceTelemetry:
      type: object
      required: [metric, resolution, points]
      properties:
        metric:
          type: string
        resolution:
          type: string
          enum: [raw, 5m, 1h]
        points:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryPoint"
    CreateAlertRule:
      oneOf:
        - $ref: "#/components/schemas/CreateThresholdAlertRule"
        - $ref: "#/components/schemas/CreateDeviceOfflineAlertRule"
    CreateThresholdAlertRule:
      type: object
      additionalProperties: false
      required: [type, name, deviceId, metric, operator, threshold]
      properties:
        type:
          type: string
          enum: [threshold]
        name:
          type: string
          minLength: 1
          maxLength: 200
        deviceId:
          type: string
          format: uuid
        metric:
          type: string
          minLength: 1
          maxLength: 128
        operator:
          type: string
          enum: [gt, gte, lt, lte]
        threshold:
          type: number
        durationSeconds:
          type: integer
          minimum: 0
          maximum: 86400
          default: 0
        hysteresis:
          type: number
          minimum: 0
          default: 0
        cooldownSeconds:
          type: integer
          minimum: 0
          maximum: 86400
          default: 0
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
    CreateDeviceOfflineAlertRule:
      type: object
      additionalProperties: false
      required: [type, name, deviceId]
      properties:
        type:
          type: string
          enum: [device_offline]
        name:
          type: string
          minLength: 1
          maxLength: 200
        deviceId:
          type: string
          format: uuid
        cooldownSeconds:
          type: integer
          minimum: 0
          maximum: 86400
          default: 0
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
    AlertRule:
      type: object
      required: [id, name, type, deviceId, metric, operator, threshold, durationSeconds, hysteresis, cooldownSeconds, severity, enabled, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [threshold, device_offline]
        deviceId:
          type: string
          format: uuid
        metric:
          type: string
        operator:
          type: string
          enum: [gt, gte, lt, lte]
        threshold:
          type: number
        durationSeconds:
          type: integer
        hysteresis:
          type: number
        cooldownSeconds:
          type: integer
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Alert:
      type: object
      required: [id, ruleId, deviceId, metric, observedValue, observedAt, message, severity, state, openedAt, resolvedAt]
      properties:
        id:
          type: string
          format: uuid
        ruleId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        metric:
          type: string
        observedValue:
          type: number
        observedAt:
          type: string
          format: date-time
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        state:
          type: string
          enum: [open, acknowledged, resolved, silenced]
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, "null"]
          format: date-time